requirements.txt
requests==2.26.0
beautifulsoup4==4.9.3
python-docx==0.8.11
